backend/requirements.txt
django
dj-database-url
psycopg2
django-allauth
djangorestframework
django-debug-toolbar
drf-flex-fields